Node api handlers and backend modules
