+++ tests/src/nodes/nomos.rs
@@ -34,7 +34,7 @@ use nomos_mempool::MempoolMetrics;
 #[cfg(feature = "mixnet")]
 use nomos_network::backends::libp2p::mixnet::MixnetConfig;
 use nomos_network::{backends::libp2p::Libp2pConfig, NetworkConfig};
-use nomos_node::{api::AxumBackendSettings, Config, Tx};
+use nomos_node::{api::backend::AxumBackendSettings, Config, Tx};
 use nomos_storage::backends::rocksdb::RocksBackendSettings;
 use once_cell::sync::Lazy;
 use rand::{thread_rng, Rng};
